+++ packages/rb-windowing/src/index.ts
@@ -5,3 +5,4 @@
 export * from './types';
 export * from './store';
 export * from './focusTrap';
+export * from './invariants';